feat: add separate MainSearchComponent
